+++ ui/src/features/dataset/gallery/gallery.component.tsx
@@ -1,10 +1,14 @@
 // Copyright (C) 2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-import { AriaComponentsListBox, GridLayout, ListBoxItem, Size, View, Virtualizer } from '@geti/ui';
+import { useState } from 'react';
 
+import { AriaComponentsListBox, DialogContainer, GridLayout, ListBoxItem, Size, View, Virtualizer } from '@geti/ui';
+
+import thumbnailUrl from '../../../assets/mocked-project-thumbnail.png';
 import { useSelectedData } from '../../../routes/dataset/provider';
 import { CheckboxInput } from '../checkbox-input';
+import { MediaPreview } from '../media-preview/media-preview.component';
 import { response } from '../mock-response';
 import { AnnotationStateIcon } from './annotation-state-icon.component';
 import { MediaItem } from './media-item.component';
@@ -17,7 +21,23 @@ const layoutOptions = {
     preserveAspectRatio: true,
 };
 
+type Item = (typeof response.items)[number];
+
+type MediaThumbnailProps = {
+    onDoubleClick: () => void;
+    url: string;
+    alt: string;
+};
+const MediaThumbnail = ({ onDoubleClick, url, alt }: MediaThumbnailProps) => {
+    return (
+        <div onDoubleClick={onDoubleClick}>
+            <img src={url} alt={alt} style={{ objectFit: 'cover', width: '100%', height: '100%' }} />
+        </div>
+    );
+};
+
 export const Gallery = () => {
+    const [selectedMediaItem, setSelectedMediaItem] = useState<null | Item>(null);
     const { selectedKeys, mediaState, setSelectedKeys } = useSelectedData();
     const isSetSelectedKeys = selectedKeys instanceof Set;
 
@@ -29,7 +49,7 @@ export const Gallery = () => {
                     aria-label='data-collection-grid'
                     className={classes.container}
                     selectedKeys={selectedKeys}
-                    selectionMode='multiple'
+                    selectionMode={'multiple'}
                     onSelectionChange={setSelectedKeys}
                 >
                     {response.items.map((item) => (
@@ -42,7 +62,13 @@ export const Gallery = () => {
                             data-rejected={mediaState.get(item.id) === 'rejected'}
                         >
                             <MediaItem
-                                item={item}
+                                contentElement={() => (
+                                    <MediaThumbnail
+                                        onDoubleClick={() => setSelectedMediaItem(item)}
+                                        url={thumbnailUrl}
+                                        alt={item.original_name}
+                                    />
+                                )}
                                 topRightElement={() => <AnnotationStateIcon state={mediaState.get(item.id)} />}
                                 topLeftElement={() => (
                                     <CheckboxInput
@@ -56,6 +82,12 @@ export const Gallery = () => {
                     ))}
                 </AriaComponentsListBox>
             </Virtualizer>
+
+            <DialogContainer onDismiss={() => setSelectedMediaItem(null)}>
+                {selectedMediaItem !== null && (
+                    <MediaPreview mediaItem={selectedMediaItem} close={() => setSelectedMediaItem(null)} />
+                )}
+            </DialogContainer>
         </View>
     );
 };